+++ lib/generators/esquema/install/install_generator.rb
@@ -4,6 +4,7 @@ require "rails/generators"
 
 module Esquema
   module Generators
+    # This generator is responsible for installing the Esquema gem.
     class InstallGenerator < Rails::Generators::Base
       source_root File.expand_path("templates", __dir__)